Organizing the code by modules
